Move send_recv_core server/client implementations into classes
